--- a/lib/benchmark/ips/report.rb
+++ b/lib/benchmark/ips/report.rb
@@ -68,8 +68,14 @@ module Benchmark
         # percentage of standard deviation, iterations in runtime.
         # @return [String] Left justified body.
         def body
-          left = "%s (±%4.1f%%) i/s" % [Helpers.scale(ips), stddev_percentage]
-          left.ljust(20) + (" - %s in %10.3fs" % [Helpers.scale(@iterations), runtime])
+          case Benchmark::IPS.options[:format]
+          when :human
+            left = "%s (±%4.1f%%) i/s" % [Helpers.scale(ips), stddev_percentage]
+            left.ljust(20) + (" - %s in %10.3fs" % [Helpers.scale(@iterations), runtime])
+          else
+            left = "%10.1f (±%.1f%%) i/s" % [ips, stddev_percentage]
+            left.ljust(20) + (" - %10d in %10.6fs" % [@iterations, runtime])
+          end
         end
 
         # Return header with padding if +@label+ is < length of 20.