@krainboltgreene 2015-04-03
The dot notation for the numbers is really throwing me off, but I also can't seem to convert to raw. I've tried passing it to the block object, as a argument to the #ips method, and more. How am I supposed to not use this new format for numbers?