@krainboltgreene 2015-04-03
Not only can I not find the way to set it, I'm fairly sure there is *no* way to set this value. There are no tests for this either.