@evanphx 2015-04-08
@krainboltgreene You can set it back to raw by adding `Benchmark::IPS.options[:format] = :raw`